+++ saleor/graphql/order/tests/queries/test_order_line.py
@@ -6,6 +6,7 @@ from django.core.files import File
 from prices import Money, TaxedMoney
 
 from .....core.prices import quantize_price
+from .....discount import DiscountType, DiscountValueType
 from .....order import OrderStatus
 from .....order.interface import OrderTaxedPricesData
 from .....thumbnail.models import Thumbnail
@@ -760,3 +761,60 @@ def test_order_query_undiscounted_prices_no_tax(
     first_order_data_line_price = order_data["lines"][0]["undiscountedUnitPrice"]
     assert first_order_data_line_price["net"]["amount"] == line.unit_price.net.amount
     assert first_order_data_line_price["gross"]["amount"] == line.unit_price.net.amount
+
+
+def test_order_line_returns_discount_object(
+    staff_api_client, order_with_lines, permission_group_all_perms_all_channels
+):
+    # given
+    expected_amount = Decimal("6")
+    expected_reason = "test"
+
+    line = order_with_lines.lines.first()
+    line.discounts.create(
+        type=DiscountType.MANUAL,
+        value_type=DiscountValueType.FIXED,
+        value=expected_amount,
+        amount_value=expected_amount,
+        currency=line.currency,
+        reason=expected_reason,
+    )
+
+    query = """
+    query OrderQuery($id: ID) {
+      order(id: $id) {
+        lines {
+          discounts{
+            valueType
+            value
+            reason
+            unit{
+                amount
+            }
+            total{
+                amount
+            }
+          }
+        }
+      }
+    }
+    """
+
+    permission_group_all_perms_all_channels.user_set.add(staff_api_client.user)
+
+    # when
+    response = staff_api_client.post_graphql(
+        query, variables={"id": to_global_id_or_none(order_with_lines)}
+    )
+
+    # then
+    content = get_graphql_content(response)
+    line_discounts = content["data"]["order"]["lines"][0]["discounts"]
+
+    assert line_discounts
+    line_discount = line_discounts[0]
+    assert line_discount["valueType"] == DiscountValueType.FIXED.upper()
+    assert line_discount["reason"] == expected_reason
+    assert line_discount["value"] == expected_amount
+    assert line_discount["unit"]["amount"] == expected_amount / line.quantity
+    assert line_discount["total"]["amount"] == expected_amount